pyproject.toml
[project]
name = "berlinger"
version = "0.1.0"
description = "Parser for Berlinger Fridge-tag 2 export files"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
    "pydantic>=2.0",
    "httpx>=0.27",
    "typer>=0.12",
    "python-dotenv>=1.0",
]

[project.scripts]
fridgetag-json = "berlinger.fridgetag_json:main"
fridgetag-dhis2 = "berlinger.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/berlinger"]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
fixable = ["ALL"]
select = ["E", "W", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
